requirements.txt
gradio>=4.0.0
PyGithub>=2.1.1
google-generativeai>=0.3.0
elevenlabs>=0.2.0
